+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-03-09T13:01:09.719Z")
+    created: Optional[datetime] = parse_datetime("2026-03-09T14:37:14.998Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")