+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK9B05YE5WNPE1WCA538A96Q"
+        "01KK9GG463Y7MTVNR34T4SM44E"
     )
     r"""The unique identifier of the data source"""