+++ packages/orq-rc/RELEASES.md
@@ -6389,3 +6389,13 @@ Based on:
 - [python v4.5.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.41 - packages/orq-rc
+
+## 2026-03-09 14:39:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.749.1 (2.855.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.42 - packages/orq-rc